fix(eg): repair the build and complete the KnowledgeStream policy lease
epistemic-graph did not compile, on main or on the candidate base. Three
controlled cargo checks established the baseline: main (e1463cf1) failed with 12
errors, and a99b1937 added 16 more. Both root causes are decomposition commits
that were never built.

81191bd8 ("centralize wave-3 compute and audio clones") introduced
modality_contract_runtime_hooks! and invoked it at item position with
self-bearing expr fragments, which cannot resolve outside a method body. The
macro now takes :path parameters and calls them inside its own generated
bodies, with each crate supplying free functions carrying the original logic.

c3d83fb8 ("harden semantic stores and split facades") shipped consumer code for
types it never wrote: git log --all -S proves EmbeddingSpaceRef, StampedVector,
MAX_EMBEDDING_DIMENSIONS, MAX_MAINTAINED_ANN_DIMENSIONS and SemanticQueryError
existed in no commit. They are authored here in a new serde-only
eg-types/embedding.rs, with duplicate inherent dim()/row_ids() removed and
cross-module visibility corrected.

a99b1937's own feature was also unfinished: GraphPolicyLease and
GraphPolicyCoarseAcl existed nowhere, so from_verified_with_lease was
uncallable and dispatch.rs still called the removed two-argument
from_verified. The subsystem is implemented here against a contract that
passed two adversarial security reviews: a non-constructible, non-Clone
GraphPolicyLease; PolicySnapshot sourcing version from the existing
security-control counter through a fail-closed RbacPolicyStore::current_version
default; a canonical digest that is never save()'s non-canonical state_digest;
a closed PolicyAwareQuery enum so no future Method variant can silently inherit
stream authorization; and a lease-filtered read path that bypasses
FilteredViewCache and ResultCache entirely, since on a cache hit filter_view is
skipped and no RBAC mutation invalidates either cache.

Minting is gated by a MintAuthorization capability token with a private field
whose sole constructor verifies an HMAC over canonically encoded claims in
constant time, so a bare RequestContextClaims literal can no longer reach the
mint. pub(crate) was not available: the caller lives in the binary crate. A
scanner gate, proven against a known-bad fixture, enforces the single
production call site. This does not defend against an in-process actor holding
auth_secret, which could bypass minting entirely; it closes accidental
misdirection and makes any new call site visible.

Also fixed en route, all pre-existing: a duplicate knowledge_stream module the
base commit's own restructure left behind; a generic helper defined inside a
trait impl in redb_backend.rs; 87 match arms across 23 functions in mutation.rs
returning bare &'static str where Option was declared, from an earlier
extraction commit; three test-target defects; and an undeclared policy_export
feature parent set, now declared in the CA-17 reserved table rather than
silenced.

cargo check -p epistemic-graph: 0 errors. --tests: 0 errors.
cargo test --lib: 1270 passed, 0 failed, 3 ignored.

+++ tests/pgwire_roundtrip.rs
@@ -892,7 +892,9 @@ fn simple_ids(msgs: Vec<tokio_postgres::SimpleQueryMessage>) -> Vec<String> {
 fn command_count(msgs: &[tokio_postgres::SimpleQueryMessage]) -> i64 {
     msgs.iter()
         .find_map(|m| match m {
-            tokio_postgres::SimpleQueryMessage::CommandComplete(n) => Some(*n),
+            // tokio-postgres yields the affected-row count as u64; this helper's
+            // callers compare against signed literals, so narrow once here.
+            tokio_postgres::SimpleQueryMessage::CommandComplete(n) => Some(*n as i64),
             _ => None,
         })
         .expect("INSERT CommandComplete")
